refactor: 서비스 레이어 제거 및 클래스명 수정
