+++ src/main/kotlin/lotto/Main.kt
@@ -1,22 +1,24 @@
 package lotto
 
-import lotto.service.OrderService
-import lotto.service.WinningLottoService
 import lotto.view.InputView
 import lotto.view.ResultView
 
 fun main() {
-    val orderService = OrderService()
-    val winningLottoService = WinningLottoService()
+    val lottoCreator = LottoCreator()
+    val lottoShop = LottoShop(lottoCreator)
+    val winningLottoMatcher = WinningLottoMatcher()
 
+    // 주문 생성
     val amount = InputView.getAmount()
-    val order = orderService.makeOrder(amount)
+    val order = lottoShop.makeOrder(amount)
     ResultView.printCreatedLottos(order.lottos)
 
+    // 당첨번호 및 보너스번호 생성
     val winNumberInput = InputView.getWinNumberInput()
     val bonusNumber = InputView.getBonusNumber()
-    val winNumbers = winningLottoService.createWinningLotto(winNumberInput, bonusNumber)
+    val winNumbers = lottoCreator.createWinningLotto(winNumberInput, bonusNumber)
 
-    val result = winningLottoService.checkAndGetResult(order, winNumbers)
+    // 결과 출력
+    val result = winningLottoMatcher.checkAndGetResult(order, winNumbers)
     ResultView.printResult(result)
 }
